Restructure exports
