+++ kolena/_experimental/dataset/__init__.py
@@ -11,3 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from kolena._experimental.dataset._dataset import fetch_dataset
+from kolena._experimental.dataset._dataset import register_dataset
+
+__all__ = [
+    "register_dataset",
+    "fetch_dataset",
+]
